requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
playwright
scikit-learn
pandas
numpy
pandas
numpy
google-genai
langchain
langchain-openai
langchain-core
langchain-community
fpdf
streamlit
python-dotenv
ultralytics
requests
supabase
